+++ src/main/kotlin/com/github/davenury/ucac/EventListener.kt
@@ -1,3 +1,9 @@
+
+package com.github.davenury.ucac
+
+import com.github.davenury.ucac.gpac.domain.Transaction
+
+
 typealias AdditionalActionConsensus = suspend () -> Unit
 
 enum class ConsensusTestAddon {
