+++ src/test/kotlin/com/github/davenury/ucac/api/MultiplePeersetSpec.kt
@@ -2,11 +2,7 @@ package com.github.davenury.ucac.api
 
 import com.fasterxml.jackson.module.kotlin.readValue
 import com.github.davenury.ucac.*
-import com.github.davenury.ucac.common.AddUserChange
-import com.github.davenury.ucac.common.ChangeDto
-import com.github.davenury.ucac.consensus.ratis.ChangeWithAcceptNum
-import com.github.davenury.ucac.consensus.ratis.ChangeWithAcceptNumDto
-import com.github.davenury.ucac.consensus.ratis.HistoryDto
+import com.github.davenury.ucac.common.*
 import com.github.davenury.ucac.gpac.domain.Accept
 import com.github.davenury.ucac.gpac.domain.Apply
 import com.github.davenury.ucac.utils.TestApplicationSet
@@ -51,7 +47,7 @@ class MultiplePeersetSpec {
             accept(ContentType.Application.Json)
         }
             .let { objectMapper.readValue<ChangeWithAcceptNumDto>(it) }
-            .let { ChangeWithAcceptNum(it.change.toChange(), it.acceptNum) }
+            .let { ChangeWithAcceptNum(it.changeDto.toChange(), it.acceptNum) }
 
         expect {
             that(peer2Change.change).isEqualTo(AddUserChange("userName"))
@@ -63,7 +59,7 @@ class MultiplePeersetSpec {
             accept(ContentType.Application.Json)
         }
             .let { objectMapper.readValue<ChangeWithAcceptNumDto>(it) }
-            .let { ChangeWithAcceptNum(it.change.toChange(), it.acceptNum) }
+            .let { ChangeWithAcceptNum(it.changeDto.toChange(), it.acceptNum) }
 
         expect {
             that(peer4Change.change).isEqualTo(AddUserChange("userName"))
@@ -71,7 +67,7 @@ class MultiplePeersetSpec {
 
         // and - there's only one change in history of both peersets
         askForChanges("http://${peers[0][1]}")
-            .let {
+            .let { it ->
                 expect {
                     that(it.size).isGreaterThanOrEqualTo(1)
                     that(it[0]).isEqualTo(ChangeWithAcceptNum(AddUserChange("userName"), 1))
@@ -79,7 +75,7 @@ class MultiplePeersetSpec {
             }
 
         askForChanges("http://${peers[1][0]}")
-            .let {
+            .let { it ->
                 expect {
                     that(it.size).isGreaterThanOrEqualTo(1)
                     that(it[0]).isEqualTo(ChangeWithAcceptNum(AddUserChange("userName"), 1))
@@ -108,7 +104,7 @@ class MultiplePeersetSpec {
 
         // then - transaction should not be executed
         askForChanges("http://${peers[0][2]}")
-            .let {
+            .let { it ->
                 expect {
                     that(it.size).isEqualTo(0)
                 }
@@ -140,7 +136,7 @@ class MultiplePeersetSpec {
 
         // then - transaction should not be executed
         askForChanges("http://${peers[0][1]}")
-            .let {
+            .let { it ->
                 expect {
                     that(it.size).isEqualTo(0)
                 }
@@ -160,7 +156,7 @@ class MultiplePeersetSpec {
 
         // then - transaction should be executed in every peerset
         askForChanges("http://${peers[0][1]}")
-            .let {
+            .let { it ->
                 expect {
                     that(it.size).isGreaterThanOrEqualTo(1)
                     that(it[0]).isEqualTo(ChangeWithAcceptNum(AddUserChange("userName"), 1))
@@ -168,7 +164,7 @@ class MultiplePeersetSpec {
             }
 
         askForChanges("http://${peers[1][0]}")
-            .let {
+            .let { it ->
                 expect {
                     that(it.size).isGreaterThanOrEqualTo(1)
                     that(it[0]).isEqualTo(ChangeWithAcceptNum(AddUserChange("userName"), 1))
@@ -234,7 +230,7 @@ class MultiplePeersetSpec {
 
         peers.flatten().forEach {
             askForChanges("http://$it")
-                .let {
+                .let { it ->
                     expect {
                         that(it.size).isGreaterThanOrEqualTo(1)
                         that(it[0].change).isEqualTo(AddUserChange("userName"))
@@ -308,7 +304,7 @@ class MultiplePeersetSpec {
 
             peers.flatten().forEach {
                 askForChanges("http://$it")
-                    .let {
+                    .let { it ->
                         expect {
                             that(it.size).isGreaterThanOrEqualTo(1)
                             that(it[0].change).isEqualTo(AddUserChange("userName"))
@@ -332,7 +328,7 @@ class MultiplePeersetSpec {
             contentType(ContentType.Application.Json)
             accept(ContentType.Application.Json)
         }.let { objectMapper.readValue<HistoryDto>(it) }
-            .changes.map { ChangeWithAcceptNum(it.change.toChange(), it.acceptNum) }
+            .changes.map { ChangeWithAcceptNum(it.changeDto.toChange(), it.acceptNum) }
 
     private val changeDto = ChangeDto(
         mapOf(
