+++ src/test/kotlin/com/github/davenury/ucac/infrastructure/InMemoryHistoryManagementSpec.kt
@@ -1,11 +1,12 @@
 package com.github.davenury.ucac.infrastructure
 
 import com.github.davenury.ucac.common.AddRelationChange
+import com.github.davenury.ucac.common.ChangeWithAcceptNum
 import com.github.davenury.ucac.common.InMemoryHistoryManagement
 import com.github.davenury.ucac.consensus.raft.domain.ConsensusFailure
 import com.github.davenury.ucac.consensus.raft.domain.ConsensusSuccess
-import com.github.davenury.ucac.consensus.ratis.ChangeWithAcceptNum
 import com.github.davenury.ucac.utils.DummyConsensusProtocol
+import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import strikt.api.expectThat
