+++ src/test/kotlin/com/github/davenury/ucac/utils/DummyConsensusProtocol.kt
@@ -1,16 +1,16 @@
 package com.github.davenury.ucac.utils
 
 import com.github.davenury.ucac.common.Change
+import com.github.davenury.ucac.common.ChangeWithAcceptNum
 import com.github.davenury.ucac.consensus.raft.domain.ConsensusProtocol
 import com.github.davenury.ucac.consensus.raft.domain.ConsensusResult
 import com.github.davenury.ucac.consensus.raft.domain.ConsensusSuccess
-import com.github.davenury.ucac.consensus.ratis.ChangeWithAcceptNum
 
 
 object DummyConsensusProtocol : ConsensusProtocol<Change, MutableList<ChangeWithAcceptNum>> {
     private var response: ConsensusResult = ConsensusSuccess
 
-    override fun proposeChange(change: Change, acceptNum: Int?): ConsensusResult = response
+    override suspend fun proposeChange(change: Change, acceptNum: Int?): ConsensusResult = response
 
 
     fun setResponse(response: ConsensusResult) {
